feat: break sync code between partial class
