+++ src/NextAudio.Matroska/Internal/BlockReader.cs
@@ -62,7 +62,7 @@ internal static class BlockReader
                 break;
         }
 
-        return new MatroskaBlockResult(new MatroskaBlock(trackNumber, lacingType, frameCount, frameSizes), position);
+        return new MatroskaBlockResult(new MatroskaBlock(blockElement, trackNumber, lacingType, frameCount, frameSizes), position);
     }
 
     public static async ValueTask<MatroskaBlockResult> ReadBlockAsync(AudioStream stream, Memory<byte> buffer, MatroskaElement blockElement, long position, ulong selectedTrackNumber, ILogger logger)
@@ -118,7 +118,7 @@ internal static class BlockReader
                 break;
         }
 
-        return new MatroskaBlockResult(new MatroskaBlock(trackNumber, lacingType, frameCount, frameSizes), position);
+        return new MatroskaBlockResult(new MatroskaBlock(blockElement, trackNumber, lacingType, frameCount, frameSizes), position);
     }
 
     private static long ParseXiphLacing(AudioStream stream, Span<byte> buffer, long position, MatroskaElement blockElement, Span<int> frameSizes)
